+++ src/components/pages/AppSubscription/AppSubscriptionDetailOverlay/index.tsx
@@ -41,7 +41,6 @@ import {
 } from 'features/appSubscription/appSubscriptionApiSlice'
 import ReleaseStepper from 'components/shared/basic/ReleaseProcess/stepper'
 import { SubscriptionStatus } from 'features/apps/types'
-import UserService from 'services/UserService'
 import { ROLES } from 'types/Constants'
 import { useState } from 'react'
 import { SuccessErrorType } from 'features/admin/appuserApiSlice'
@@ -50,6 +49,7 @@ import { SubscriptionTypes } from 'components/shared/templates/Subscription'
 import { useFetchServiceSubDetailQuery } from 'features/serviceSubscription/serviceSubscriptionApiSlice'
 import HelpOutlineIcon from '@mui/icons-material/HelpOutline'
 import { Link } from 'react-router-dom'
+import { userHasPortalRole } from 'services/AccessService'
 
 interface AppSubscriptionDetailProps {
   openDialog: boolean
@@ -215,7 +215,7 @@ const AppSubscriptionDetailOverlay = ({
   const renderTenantUrl = (url: string) => {
     if (
       isAppSubscription &&
-      UserService.hasRole(ROLES.APPSTORE_EDIT) &&
+      userHasPortalRole(ROLES.APPSTORE_EDIT) &&
       data?.offerSubscriptionStatus === SubscriptionStatus.ACTIVE
     ) {
       return (