+++ src/services/AccessService.tsx
@@ -18,8 +18,13 @@
  * SPDX-License-Identifier: Apache-2.0
  ********************************************************************************/
 
-import type { IAction, IOverlay, IPage, Tree } from 'types/MainTypes'
-import UserService from './UserService'
+import type {
+  IAction,
+  IOverlay,
+  IPage,
+  RestrictedItem,
+  Tree,
+} from 'types/MainTypes'
 import { Route } from 'react-router-dom'
 import AppInfo from 'components/overlays/AppInfo'
 import AddBPN from 'components/overlays/AddBPN'
@@ -71,16 +76,42 @@ import CompanyCertificateDetails from 'components/overlays/CompanyCertificateDet
 import DeleteCompanyCertificateConfirmationOverlay from 'components/overlays/CompanyCertificateDetails/DeleteCompanyCertificateConfirmationOverlay'
 import { DisableManagedIDP } from 'components/overlays/EnableIDP/DisableManagedIdp'
 import { DeleteManagedIDP } from 'components/overlays/IDPDelete/DeleteManagedIdp'
+import type { KeycloakResourceAccess } from 'keycloak-js'
+import { intersectAccess } from 'utils/dataUtils'
+import { getAccess } from './UserService'
+import {
+  getClientId,
+  getClientIdMiw,
+  getClientIdSemantic,
+  getClientIdSsiCredential,
+} from './EnvironmentService'
 
 let pageMap: { [page: string]: IPage }
 let actionMap: { [action: string]: IAction }
 let overlayMap: { [overlay: string]: IOverlay }
 
-const checkAccess = (element: { name: string; role?: string }): boolean => {
-  if (!element) return false // page doesn't exist
-  const role = element.role
-  if (!role) return true // no permission required for this page
-  return UserService.hasRole(role) // check if user has the required permission
+export const userHasAccess = (required: KeycloakResourceAccess): boolean =>
+  Object.keys(intersectAccess(required, getAccess())).length > 0
+export const userHasClientRole = (
+  client: string,
+  roles: string | Array<string>
+): boolean =>
+  userHasAccess({ [client]: { roles: Array.isArray(roles) ? roles : [roles] } })
+export const userHasPortalRole = (roles: string | Array<string>): boolean =>
+  userHasClientRole(getClientId(), roles)
+export const userHasSemanticHubRole = (
+  roles: string | Array<string>
+): boolean => userHasClientRole(getClientIdSemantic(), roles)
+export const userHasMiwRole = (roles: string | Array<string>): boolean =>
+  userHasClientRole(getClientIdMiw(), roles)
+export const userHasSsiCredentialRole = (
+  roles: string | Array<string>
+): boolean => userHasClientRole(getClientIdSsiCredential(), roles)
+
+const checkAccess = (element: RestrictedItem): boolean => {
+  if (!element) return false // item doesn't exist
+  if (!element.allowTo) return true // no permission required for this item
+  return element.allowTo() // check if permission is granted
 }
 
 export const hasAccess = (page: string) => checkAccess(pageMap[page])
@@ -94,9 +125,7 @@ export const hasAccessOverlay = (overlay: string) =>
 const accessToMenu = (menu: string[]) =>
   menu.filter((page: string) => hasAccess(page))
 
-// Add an ESLint exception until there is a solution
-// eslint-disable-next-line
-const accessToMenuTree = (menu: Tree[] | undefined): any =>
+const accessToMenuTree = (menu: Tree[] | undefined): Tree[] | undefined =>
   menu
     ?.filter((item: Tree) => hasAccess(item.name))
     .map((item: Tree) => ({